+++ pystarport/tests/test_expansion/test_expansion.py
@@ -1,34 +1,45 @@
 import os
 from pathlib import Path
 
+import pytest
 import yaml
 from deepdiff import DeepDiff
 
-from pystarport.expansion import expand_yaml
+from pystarport.expansion import expand_jsonnet, expand_yaml
 
 
-def test_expansion():
+def _get_base_config():
+    return yaml.safe_load(open(Path(__file__).parent / "base.yaml"))
+
+
+@pytest.mark.parametrize(
+    "type, func",
+    [(".yaml", expand_yaml), (".jsonnet", expand_jsonnet)],
+)
+def test_expansion(type, func):
     parent = Path(__file__).parent
-    base = parent / "base.yaml"
-    cronos_has_dotenv = parent / "cronos_has_dotenv.yaml"
-    cronos_no_dotenv = parent / "cronos_no_dotenv.yaml"
-    cronos_has_posix_no_dotenv = parent / "cronos_has_posix_no_dotenv.yaml"
-
-    baseConfig = yaml.safe_load(open(base))
+    cronos_has_dotenv = parent / ("cronos_has_dotenv" + type)
+    cronos_no_dotenv = parent / ("cronos_no_dotenv" + type)
+    cronos_has_posix_no_dotenv = parent / ("cronos_no_dotenv" + type)
+    baseConfig = _get_base_config()
     # `expand_yaml` is backward compatible, not expanded, and no diff
-    assert baseConfig == expand_yaml(cronos_no_dotenv, None)
+    config = func(cronos_no_dotenv, None)
+    assert baseConfig == config
 
     # `expand_yaml` is expanded but no diff
+    config = func(cronos_has_dotenv, None)
     assert not DeepDiff(
         baseConfig,
-        expand_yaml(cronos_has_dotenv, None),
+        config,
         ignore_order=True,
     )
 
     # overriding dotenv with relative path is expanded and has diff)
+    dotenv = "dotenv1"
+    config = func(cronos_has_dotenv, dotenv)
     assert DeepDiff(
         baseConfig,
-        expand_yaml(cronos_has_dotenv, "dotenv1"),
+        config,
         ignore_order=True,
     ) == {
         "values_changed": {
@@ -42,9 +53,11 @@ def test_expansion():
     }
 
     # overriding dotenv with absolute path is expanded and has diff
+    dotenv = os.path.abspath("test_expansion/dotenv1")
+    config = func(cronos_has_dotenv, dotenv)
     assert DeepDiff(
         baseConfig,
-        expand_yaml(cronos_has_dotenv, os.path.abspath("test_expansion/dotenv1")),
+        config,
         ignore_order=True,
     ) == {
         "values_changed": {
@@ -58,10 +71,10 @@ def test_expansion():
     }
 
     # overriding dotenv with absolute path is expanded and no diff
+    dotenv = os.path.abspath("test_expansion/dotenv")
+    config = func(cronos_has_posix_no_dotenv, dotenv)
     assert not DeepDiff(
         baseConfig,
-        expand_yaml(
-            cronos_has_posix_no_dotenv, os.path.abspath("test_expansion/dotenv")
-        ),
+        config,
         ignore_order=True,
     )